event: New event data serialization system
This is a pretty big change which more-or-less needed to be made all at
once. Rather than directly implementing ReadFrom and WriteTo, event
types now implement SerializeEvent and DeserializeEvent. This is
inspired by serde and it decouples the underlying transport from the
event data so that we can e.g. switch opcode tables based on region or
translate commands to an assembly language format.