cli: Experimental script disassembler
This introduces a new assembly language representation for script files
which should theoretically be easy to edit and reassemble into binary
script data. The language is structured such that it directly
corresponds to how instructions appear in the stage file, which is why
I've named it an "assembly language."

You can use the `script disassemble` and `script disassemble-all`
commands to disassemble stage file scripts, and there is a simple Visual
Studio Code extension in unplug-vscode/ which provides syntax
highlighting.

The language syntax and structure is subject to change; right now this
is basically just a proof-of-concept. The ability to disassemble and
then reassemble scripts is the main feature planned for the 0.5 release.