common: Rewrite of the Text struct
This is a pretty big refactor which allows `Text` to have different data
formats, including a `Vec<u8>`, `CString`, or byte array. The rules for
each format are enforced at runtime and it is impossible for `Text` to
ever contain interior NULs or invalid data.

We also now store more information about the text in the type, so we can
do I/O more easily without needing the `fixed_string` functions anymore
or needing to use `CString` directly.