common: I/O extension traits for read_struct() and write_struct()
These traits extend `Read` and `Write` so that it's easier to read/write
types that implement `ReadFrom`/`WriteTo`. In particular,
`read_struct()` lets the compiler do type inference so we don't have to
type out the path to the `read_from()` call.

Technically this can be used to write types which aren't structs (e.g.
arrays), but for most use cases "struct" is clear.